CSSPreloadScanner should not depend on HTMLToken https://bugs.webkit.org/show_bug.cgi?id=109742
Reviewed by Eric Seidel.

There's no need for the CSSPreloadScanner to depend on HTMLToken. On
the background thread, we'll likely want to use a CompactHTMLToken for
preload scanning, so this dependency is problematic. This patch also
teaches the CSSPreloadScanner how to scan LChars.

* html/parser/CSSPreloadScanner.cpp:
(WebCore::CSSPreloadScanner::~CSSPreloadScanner):
(WebCore):
(WebCore::CSSPreloadScanner::scan):
* html/parser/CSSPreloadScanner.h:
(CSSPreloadScanner):
* html/parser/HTMLPreloadScanner.cpp:
(WebCore::HTMLPreloadScanner::processToken):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@142803 268f45cc-cd09-0410-ab3c-d52691b4dbfc