[GTK] Errors when building WebKit2 with Clang https://bugs.webkit.org/show_bug.cgi?id=109603
Reviewed by Alexey Proskuryakov.

Source/WebKit2: 

* UIProcess/API/gtk/WebKitUIClient.cpp:
(setWindowFrame): Cast the position parameters to the integer type when constructing the GdkRectangle.
* UIProcess/Authentication/WebCredential.cpp: Add an empty implementation of the newly specified destructor.
* UIProcess/Authentication/WebCredential.h: Work around the incomplete WebCertificateInfo type errors thrown in the inline
RefPtr destructor of the m_certificateInfo class member by defining a destructor in the class.
* WebProcess/WebCoreSupport/WebEditorClient.h: The GTK-specific methods are not virtual and as such cannot be overriden.
The OVERRIDE keywords are thus unnecessary.

Tools: 

* WebKitTestRunner/gtk/PlatformWebViewGtk.cpp:
(WTR::PlatformWebView::resizeTo): Cast the width and height parameters to the integer
type when constructing the GtkAllocation.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@142908 268f45cc-cd09-0410-ab3c-d52691b4dbfc