drop-shadow filter with overflow:hidden child misbehaves https://bugs.webkit.org/show_bug.cgi?id=109783
Source/WebCore:

Reviewed by Dean Jackson.

The change in r112745 was not sufficient; it failed to account
for descendant layers that needed to not clipping to avoid artefacts
with filters like drop-shadow.

Test: css3/filters/filter-repaint-shadow-layer-child.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::paintLayerContents): Remove the useClipRect bool.
Replace it with a clipToDirtyRect member on the LayerPaintingInfo, which
gets passed to descendants. Remove some "Restore the clip" comments that added
nothing.
* rendering/RenderLayer.h:
(WebCore::RenderLayer::LayerPaintingInfo::LayerPaintingInfo):
(LayerPaintingInfo):

LayoutTests:

Reviewed by Dean Jackson.

* css3/filters/filter-repaint-shadow-layer-child-expected.html: Added.
* css3/filters/filter-repaint-shadow-layer-child.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@143070 268f45cc-cd09-0410-ab3c-d52691b4dbfc