Remove unnecessary includes for Platform.h from headers https://bugs.webkit.org/show_bug.cgi?id=110300
Patch by Laszlo Gombos <l.gombos@samsung.com> on 2013-02-20
Reviewed by Alexey Proskuryakov.

Source/WebCore:

wtf/Platform.h only need to be included in a few places directly
as it is included in config.h. Remove the unnecessay duplicate as
it can be confusing.

No new tests, no change in functionality.

* platform/SecureTextInput.h:
* platform/audio/Biquad.h:
* platform/audio/FFTFrame.h:
* platform/graphics/cpu/arm/filters/FEBlendNEON.h:
* platform/graphics/cpu/arm/filters/NEONHelpers.h:
* platform/graphics/filters/FELighting.h:

Source/WebKit2:

wtf/Platform.h only need to be included in a few places directly
as it is included in config.h. Remove the unnecessay duplicate as
it can be confusing.

* Shared/PrintInfo.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@143499 268f45cc-cd09-0410-ab3c-d52691b4dbfc