Line layout (but not pref widths) double-counts word spacing when between inlines https://bugs.webkit.org/show_bug.cgi?id=110381
Reviewed by Emil A Eklund.

Source/WebCore:

When a word break with a space occurs between inlines and word-spacing is non-zero,
line layout incorrectly committed the word-spacing value to the line twice, causing
wrapping in inline-blocks. To fix this, clear the word-spacing value when between
words and ignoring spaces, as we will already have added the word-spacing to the
line.

Test: fast/text/word-space-between-inlines.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlock::LineBreaker::nextSegmentBreak):

LayoutTests:

* fast/text/word-space-between-inlines-expected.html: Added.
* fast/text/word-space-between-inlines.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@143520 268f45cc-cd09-0410-ab3c-d52691b4dbfc