Threaded HTML Parser fails fast/dom/Document/readystate.html https://bugs.webkit.org/show_bug.cgi?id=110529
Reviewed by Eric Seidel.

Source/WebCore:

We need to create the tokenizer even when we finish a script-created
parser. Most of the time, the tokenizer has already been created at
this point because it is created by calls to document.write.

Test: fast/parser/body-should-exist-after-open-close.html

* html/parser/HTMLDocumentParser.cpp:
(WebCore::HTMLDocumentParser::finish):

LayoutTests:

This test is a reduction of the original issue.

* fast/parser/body-should-exist-after-open-close-expected.txt: Added.
* fast/parser/body-should-exist-after-open-close.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@143670 268f45cc-cd09-0410-ab3c-d52691b4dbfc