REGRESSION(r143654): some fast/js test crashes on 32 bit build https://bugs.webkit.org/show_bug.cgi?id=110590
Reviewed by Mark Hahnenberg.
        
In compileValueToInt32, the refactoring in r143654 undid one of the fixes from
r143314 due to a merge goof.
        
In speculateNumber, we were simply forgetting to indicate that we need a
ManualOperandSpeculation on a JSValueOperand. ManualOperandSpeculation should
be passed whenever you will be performing the type checks yourself rather than
using the operand class to do it for you.

* dfg/DFGSpeculativeJIT.cpp:
(JSC::DFG::SpeculativeJIT::compileValueToInt32):
(JSC::DFG::SpeculativeJIT::speculateNumber):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@143800 268f45cc-cd09-0410-ab3c-d52691b4dbfc