Web Inspector: Adding existing key in DOMStorageItemsView leaves it inconsistent state https://bugs.webkit.org/show_bug.cgi?id=111061
Reviewed by Alexander Pavlov.

The view must be checked for all the rows with the matching keys. Only the first one would be updated
with the new value while the others must be removed from the grid. Also in case there are various key/value
pairs, the changed node should be selected and revealed in the grid.

No new tests as UI related minor change.

* inspector/front-end/DOMStorageItemsView.js:
(WebInspector.DOMStorageItemsView.prototype._domStorageItemUpdated):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@144317 268f45cc-cd09-0410-ab3c-d52691b4dbfc