REGRESSION: fast/border/border-fit-2.html needs updating https://bugs.webkit.org/show_bug.cgi?id=111776
Reviewed by Simon Fraser.

Source/WebCore: 

This test is incorrectly shrinking the border image now.
We need to apply some clamping to the border-fit like the
old code did. The pixel results *are* still changing though,
since a layout-time shrinkage will not result in right-aligned
boxes in an LTR block when the left edge gets chopped. We'll
have to see if this behavioral change ends up being a problem,
but for now we'll rebaseline and assume it will be ok.

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::fitBorderToLinesIfNeeded):

LayoutTests: 

* platform/mac/fast/borders/border-fit-2-expected.png:
* platform/mac/fast/borders/border-fit-2-expected.txt:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145139 268f45cc-cd09-0410-ab3c-d52691b4dbfc