Layout Test fast/dom/HTMLMeterElement/meter-element-crash.html is failing https://bugs.webkit.org/show_bug.cgi?id=88131
Reviewed by Tony Chang.

This test was calling setTimeout(..., 0) in a <script> block before the end of the document.
The timeout sometimes fired before parsing the rest of the document and sometimes after,
depending on when the parser yielded. This moves the <script> block to the end of the document
so it always runs after parsing the rest of the test, which is the more common and expected
behavior.

* fast/dom/HTMLMeterElement/meter-element-crash.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145263 268f45cc-cd09-0410-ab3c-d52691b4dbfc