[GTK] WebKit2-only build fails https://bugs.webkit.org/show_bug.cgi?id=112033
Reviewed by Gustavo Noronha Silva.

Source/WebKit2: 

* GNUmakefile.am: Introduce the artifical dependency of libwebkit2gtk on libwebkitgtk only
if building WebKit1 as well. Similarly with the WebKitPluginProcess, depend on libwebkitgtk
only if building WebKit1, the dependency is otherwise not necessary.

Tools: 

* GNUmakefile.am: Split the list of files the docs-build.stamp target depends on into the
generic part, WebKit1-specific part and WebKit2-specific part.
* TestWebKitAPI/GNUmakefile.am: The TestGtk executable should not link agains libwebkitgtk
as it is not WebKit1-specific, but should rather link to libWebCore.la and libWebCoreGtk.la
directly.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145395 268f45cc-cd09-0410-ab3c-d52691b4dbfc