Don't send events to plugins that are being snapshotted https://bugs.webkit.org/show_bug.cgi?id=112215 <rdar://problem/13386224>
Reviewed by Tim Horton.

If a plugin is in the process of being snapshotted, do not forward it
any events. This avoids the case where we can accidentally capture some
custom mouse pointers, as well as trigger a plugin menu.

The next step will be to detect if the event should stop the snapshotting
(restart the plugin).

* html/HTMLPlugInElement.cpp:
(WebCore::HTMLPlugInElement::defaultEventHandler): Return early if the plugin is snapshotting.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145645 268f45cc-cd09-0410-ab3c-d52691b4dbfc