[CSS Regions] Break after doesn't work correctly with auto-height regions https://bugs.webkit.org/show_bug.cgi?id=112241
Reviewed by David Hyatt.

Source/WebCore:

The change correctly returns the break position during the normal layout phase for
auto-height regions. The problem doesn't exist during the constrained layout phase
so it only affects the size of the auto-height regions by incorrectly positioning
the content inside the flow.

Test: fast/regions/autoheight-break-after.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::applyAfterBreak): Include the margin offset in the returned break position.

LayoutTests:

This reference test verifies the correct break positioned is computed when there are auto-height regions during
the normal layout phase.
The test includes an auto-height region, a region with a specified size and an auto-height region. The first
region ends with a paragraph with a forced break after. During the normal layout phase, if the paragraph margin
is not included in the break offset, the next paragraph is incorrectly positioned and it will result in an
incorrectly sized last region.
In the constrained layout phase, the break position is returned correctly, the paragraph is positioned further down
causing overflow in the auto-height region.

* fast/regions/autoheight-break-after-expected.html: Added.
* fast/regions/autoheight-break-after.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@145728 268f45cc-cd09-0410-ab3c-d52691b4dbfc