GenerateHashValue should be usable outside CodeGeneratorJS.pm https://bugs.webkit.org/show_bug.cgi?id=112736
Reviewed by Benjamin Poulain.

GenerateHashValue is not specific to CodeGeneratorJS.pm. Instead, the
function matches the StringHash used in WTF. This patch moves the
function into a separate Perl module so that it can be used by other
Perl scripts. For example, I plan to use this function to compute the
hash value for HTMLNames at compile time rather than at startup.

* bindings/scripts/CodeGeneratorJS.pm:
(GenerateHashTable):
* bindings/scripts/Hash.pm: Added.
(leftShift):
(GenerateHashValue):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@146253 268f45cc-cd09-0410-ab3c-d52691b4dbfc