Web Inspector: Flame Chart. Chart has to be zoomed around the mouse pointer. https://bugs.webkit.org/show_bug.cgi?id=113031
Reviewed by Yury Semikhatsky.

Overview grid is able to zoom around the mouse pointer.
So the simplest way is to translate x coordinate of the mouse pointer hovered over the chart
to x coordinate in term of overview window and pass it to the zoom function.

Minor unrelated fix: draw small border at bottom of the overview grid.

* inspector/front-end/FlameChart.js:
(WebInspector.FlameChart.prototype._onMouseWheel):
* inspector/front-end/OverviewGrid.js:
(WebInspector.OverviewGrid.prototype.zoom):
* inspector/front-end/flameChart.css:
(#flame-chart-overview-container):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@146591 268f45cc-cd09-0410-ab3c-d52691b4dbfc