[CSS Exclusions] The radius of a circle should be computed based on the shorter available dimension https://bugs.webkit.org/show_bug.cgi?id=113255
Reviewed by Julien Chaffraix.

When we decide what should be the radius of a circle, we should choose the smallest available space. For instance when the
width or height is not resolvable for the circle we should not have a radius for it. This change modifies the behavior to
use the smaller available space, so we won't render unnecessary circle shapes.

Source/WebCore: 

Test: fast/exclusions/shape-inside/shape-inside-on-nested-container-with-unresolved-height.html

* rendering/ExclusionShape.cpp:
(WebCore::ExclusionShape::createExclusionShape):

LayoutTests: 

* fast/exclusions/shape-inside/shape-inside-on-nested-container-with-unresolved-height-expected.html: Added.
* fast/exclusions/shape-inside/shape-inside-on-nested-container-with-unresolved-height.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@146938 268f45cc-cd09-0410-ab3c-d52691b4dbfc