[shadow] styleForText should consider the case where parent node has no style https://bugs.webkit.org/show_bug.cgi?id=113275
Source/WebCore:

If a text node is distributed, its parent node for rendering and
styling might have no style. In the case, styleForText should return
default style.

Reviewed by Hajime Morrita.

Test: fast/dom/shadow/text-node-distributed-crash.html

* css/StyleResolver.cpp:
(WebCore::StyleResolver::styleForText):
Check whether a parent node has any style or not. If not, return
default style. This is the same behavior as styleForElement.

LayoutTests:

Reviewed by Hajime Morrita.

* fast/dom/shadow/text-node-distributed-crash-expected.txt: Added.
* fast/dom/shadow/text-node-distributed-crash.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@146967 268f45cc-cd09-0410-ab3c-d52691b4dbfc