Support C++11 static_assert https://bugs.webkit.org/show_bug.cgi?id=113308
Reviewed by Alexey Proskuryakov.

Use the C++11 static_assert for COMPILE_ASSERT when it is available. Note that
even when compiling with C++11 support, static_assert will not always be available
since Assertions.h may also be used outside of C++.

C11 _Static_assert is enabled for GCC when it has C11 support (gcc>=4.7 -std=c11).

* wtf/Assertions.h:
* wtf/Compiler.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@146993 268f45cc-cd09-0410-ab3c-d52691b4dbfc