Prepare extend-selection-enclosing-block test to support different editing behavior https://bugs.webkit.org/show_bug.cgi?id=113424
Reviewed by Ryosuke Niwa.

We make this test Mac specific and change
extendAndLogSelectionWithinBlock() to take the platform as a
parameter. When bug 110487 is fixed this should be extended to
test the Windows editing behavior.

* editing/selection/extend-selection-enclosing-block-mac-expected.txt:
Renamed from
LayoutTests/editing/selection/extend-selection-enclosing-block-expected.txt.
* editing/selection/extend-selection-enclosing-block-mac.html:
Renamed from
LayoutTests/editing/selection/extend-selection-enclosing-block.html.
* editing/selection/resources/extend-selection.js:
(extendAndLogSelectionWithinBlock): Extend to receive platform as
a parameter. Only "mac" is supported for now.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@147146 268f45cc-cd09-0410-ab3c-d52691b4dbfc