SharedMemory release build hardening. <rdar://problem/13546860> and https://bugs.webkit.org/show_bug.cgi?id=113710
Reviewed by Anders Carlsson.

Successful creation of a memory entry with a size less than the size we requested should be treated
as a failure to create the SharedMemory object.

* Platform/mac/SharedMemoryMac.cpp:
(WebKit::SharedMemory::createFromVMBuffer): Early return if the port we get is bogus for our needs.
(WebKit::SharedMemory::createHandle): Ditto.
(WebKit::SharedMemory::create): Guarantee we pass valid input in to mach_vm_map.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@147342 268f45cc-cd09-0410-ab3c-d52691b4dbfc