Remove a couple of malloc from ExceptionBase construction https://bugs.webkit.org/show_bug.cgi?id=113681
Reviewed by Darin Adler.

* dom/ExceptionBase.cpp:
(WebCore::ExceptionBase::ExceptionBase):
When the condition is true, we were creating a new String for
description.name. The constructor had already allocated a string for
that: m_name. Use that string instead of creating a new one.

When the condition is false, we were creating a String for typeName
just to use the string operators. This is a waste of time, we can use
makeString() to invoke the string concatenation functions directly.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@147353 268f45cc-cd09-0410-ab3c-d52691b4dbfc