[BlackBerry] Create a selection from two points irregardless of the caret position. https://bugs.webkit.org/show_bug.cgi?id=113795
Patch by Nima Ghanavatian <nghanavatian@blackberry.com> on 2013-04-02
Reviewed by Rob Buis.

Internally reviewed by Mike Fenton.

PR276577
If we are to set both the base and extent of a selection, we should maintain the base upon setting the extent.
This prevents us from overwriting the previous calculation when setting extent.
Further, instead of setting and manipulating the selection, simply set up start and end VisiblePositions and
create a new directional selection so as to set the base and extent simultaneously.

* WebKitSupport/SelectionHandler.cpp:
(BlackBerry::WebKit::SelectionHandler::setSelection):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@147467 268f45cc-cd09-0410-ab3c-d52691b4dbfc