REGRESSION(r145279): Build fails while linking when MEDIA_STREAM is enabled https://bugs.webkit.org/show_bug.cgi?id=113294
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2013-04-05
Reviewed by Xan Lopez.

In r145279 NamedGetter extended attribute is added to
RTCStatsResponse.idl which makes the code generator to add
prototypes for JSRTCStatsResponse::nameGetter and
JSRTCStatsResponse::canGetItemsForName to the generated
JSRTCStatsResponse header, but they are not implemented
anywhere. We need to provide a JSRTCStatsResponseCustom file with
the implementation of both functions.

* GNUmakefile.list.am: Add JSRTCStatsResponseCustom.cpp to
compilation.
* Target.pri: Ditto.
* UseJSC.cmake: Ditto.
* WebCore.vcproj/WebCore.vcproj: Ditto.
* WebCore.vcxproj/WebCore.vcxproj: Ditto.
* WebCore.vcxproj/WebCore.vcxproj.filters: Ditto.
* bindings/js/JSRTCStatsResponseCustom.cpp: Added.
(WebCore):
(WebCore::JSRTCStatsResponse::canGetItemsForName):
(WebCore::JSRTCStatsResponse::nameGetter):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@147731 268f45cc-cd09-0410-ab3c-d52691b4dbfc