Initialize logging channels for web processes https://bugs.webkit.org/show_bug.cgi?id=114621
Reviewed by Sam Weinig.

Use InitializeWebKit2 for initializing the web and plugin processes.
It handles initializing the logging channels and it's better to have
shared core for this. For the plugin process now we always create
a RunLoop object which is not necessary with the -scanPlugin argument
but it shouldn't be a problem.

* PluginProcess/qt/PluginProcessMainQt.cpp:
(WebKit::PluginProcessMain):
* PluginProcess/unix/PluginProcessMainUnix.cpp:
(WebKit::PluginProcessMainUnix):
* WebProcess/efl/WebProcessMainEfl.cpp:
(WebKit::WebProcessMainEfl):
* WebProcess/gtk/WebProcessMainGtk.cpp:
(WebKit::WebProcessMainGtk):
* WebProcess/qt/WebProcessMainQt.cpp:
(WebKit::WebProcessMainQt):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@148519 268f45cc-cd09-0410-ab3c-d52691b4dbfc