[Dashboard] Clean up the timeline updating code, removing Chromium-specific cases https://bugs.webkit.org/show_bug.cgi?id=114756
Patch by Zan Dobersek <zandobersek@gmail.com> on 2013-04-17
Reviewed by Ryosuke Niwa.

Clean up the updating of the timeline in the timeline explorer, defaulting to the webkit.org build master when
constructing the results URL and removing the possible row addition containing the Chromium commit range that
only applied to Chromium builders in the first place. Also removes the now-redundant shouldShowWebKitRevisionsOnly
method.

* TestResultServer/static-dashboards/timeline_explorer.js:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@148631 268f45cc-cd09-0410-ab3c-d52691b4dbfc