CSS attribute selectors cause unnecessary style recalc when setting attribute to same value. <http://webkit.org/b/115116> <rdar://problem/13727709>
Reviewed by Simon Fraser.

The logic that dirties the style if there's a relevant attribute selector in the document
shouldn't run if the attribute is being overwritten with an identical value.
Move this into willModifyAttribute() instead, since we need access to both the old and the new value.

This reduces unnecessary style recalculation in Mac App Store content.

* dom/Element.cpp:
(WebCore::Element::attributeChanged):
(WebCore::Element::willModifyAttribute):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149047 268f45cc-cd09-0410-ab3c-d52691b4dbfc