Setting an inline style property to "" shouldn't cause style recalc unless the property was present. <http://webkit.org/b/115122>
Reviewed by Antti Koivisto.

StylePropertySet::setProperty() handles empty strings by removing the property entirely.
If there was no property to remove, it shouldn't claim that it changed something, as that
will cause the CSSOM wrapper to mark the element for style recalc.

A common idiom that triggers this is excessive use of { element.style.display = ''; }

* css/StylePropertySet.cpp:
(WebCore::MutableStylePropertySet::setProperty):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149059 268f45cc-cd09-0410-ab3c-d52691b4dbfc