PDFPlugin: Cursor should become an i-beam when over selectable text https://bugs.webkit.org/show_bug.cgi?id=115018 <rdar://problem/12645012>
Reviewed by Alexey Proskuryakov.

* WebProcess/Plugins/PDF/PDFPlugin.h:
Add updateCursor(), the UpdateCursorMode and HitTestResult enums, and
storage for the last PDF hit test result.
* WebProcess/Plugins/PDF/PDFPlugin.mm:
(WebKit::PDFPlugin::updateCursor):
Hit test the PDF. If there is a word under the cursor, use the I-beam cursor,
otherwise, use the default pointer cursor. This approximates PDFKit's behavior,
though does not match it precisely in the case of annotations.
Only send the cursor update to the UIProcess if it is a change, or if
we're explicitly asked to update.
(WebKit::PDFPlugin::handleMouseEvent):
If we're processing a mouse moved event, update the cursor if needed.
(WebKit::PDFPlugin::handleMouseEnterEvent):
If the cursor just moved over the PDFPlugin, force-update the cursor,
even if it matches the previously-set cursor (because something else
likely updated the cursor while it was not over the PDFPlugin).

* WebCore.exp.in: Export the I-beam and arrow cursors.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149091 268f45cc-cd09-0410-ab3c-d52691b4dbfc