[CSS Regions] Hit testing is broken for absolutely positioned regions that have overflow: hidden https://bugs.webkit.org/show_bug.cgi?id=113874
Reviewed by David Hyatt.

Source/WebCore:

Test: fast/regions/hit-test-abspos-overflow-region.html

When a region is an out-of-flow positioned object with an overflow clip, we need
to make sure that hit testing works also for cases other than foreground (content)
hit testing. This patch moves the previous hit testing code, that delegated foreground
hit testing to the region's flow thread hit testing, into the hitTestContent method,
now that RenderRegion is RenderBlock based.

* rendering/RenderRegion.cpp:
(WebCore::RenderRegion::hitTestContents):
* rendering/RenderRegion.h:
(RenderRegion):

LayoutTests:

Add test for hit-testing of border of an out-of-flow region with overflow: hidden.

* fast/regions/hit-test-abspos-overflow-region-expected.txt: Added.
* fast/regions/hit-test-abspos-overflow-region.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149168 268f45cc-cd09-0410-ab3c-d52691b4dbfc