[WK2][EFL] Build break after r149212 https://bugs.webkit.org/show_bug.cgi?id=115294
Patch by Yael Aharon <yael.aharon@intel.com> on 2013-04-26
Reviewed by Simon Fraser.
r144672 bumped up the revision of WKContextClient API, and introduced a version array
APIClientTraits<WKContextClient>::interfaceSizesByVersion .
In r149212, the version was bumped back down and the version array was emptied but not removed.
This causes a warning "array subscript is below array bounds", and in EFL the warning is
treated as error.

* Shared/APIClientTraits.cpp:
(WebKit):
* Shared/APIClientTraits.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@149230 268f45cc-cd09-0410-ab3c-d52691b4dbfc